requirements_build.txt
west==0.13.*
PyYAML==6.0
GitPython==3.1.*
joblib==1.1.0
-e git+https://github.com/antmicro/dts2repl.git@e0db6ac#egg=dts2repl
